requirements.txt
# requirements.txt for FPTU_RAG

# Core dependencies
playwright==1.40.0
python-dotenv==1.0.0
streamlit==1.29.0

# Web framework
flask==2.3.3

# Graph database
neo4j==5.15.0

# Embedding and vector search
sentence-transformers==2.2.2
torch==2.1.0
faiss-cpu==1.7.4  # Use faiss-gpu instead if you have compatible CUDA
numpy==1.24.3

# Generation model
google-generativeai==0.3.0

# Utilities
tqdm==4.66.1
uuid==1.30
pandas==2.0.3
pydantic==2.4.2

# Additional packages for processing
scikit-learn==1.3.0
nltk==3.8.1
beautifulsoup4==4.12.2
transformers==4.34.0